Implemented service worker and downloaded game file into Indexed-DB
