+++ web-client/public/index.html
@@ -58,4 +58,14 @@
     -->
 </body>
 
+<script>
+  if ("serviceWorker" in navigator) {
+    window.addEventListener("load", () => {
+      navigator.serviceWorker.register("serviceWorker.js").then(reg => {
+        console.log("We found your service worker file!", reg);
+      });
+    });
+  }
+</script>
+
 </html>
